base1: Fix and run test-journal-renderer.js
As we don't build a "lib" webpack, this test has never run during `make
check`. Move it to base1, fix it to actually work with current QUnit,
and modernize to ES6 syntax.
